+++ src/test/java/io/github/novacrypto/IntArrayPublicDerivationTests.java
@@ -21,8 +21,8 @@
 
 package io.github.novacrypto;
 
-import io.github.novacrypto.bip32.Derive;
-import io.github.novacrypto.bip32.IntArrayDerivation;
+import io.github.novacrypto.bip32.derivation.Derive;
+import io.github.novacrypto.bip32.derivation.IntArrayDerivation;
 import io.github.novacrypto.bip32.PrivateKey;
 import io.github.novacrypto.bip32.PublicKey;
 import io.github.novacrypto.bip32.networks.Bitcoin;
